+++ src/hydraulics.cpp
@@ -236,6 +236,7 @@ double Egamma(double psi, double kxylemmax, double c, double d, double psiCav = 
   double z = pow(psi/d,c);
   double* pq = incgam(h,z);
   double g = tgamma(h)*pq[0]; //Upper incomplete gamma, without the normalizing factor
+  delete[] pq;
   double E = kxylemmax*(-d/c)*g;
   if(psiCav<0.0) { //Decrease E from 0 to psiCav (avoid recursiveness!)
     if(psiCav < psi) {
@@ -243,6 +244,7 @@ double Egamma(double psi, double kxylemmax, double c, double d, double psiCav = 
     } else {
       double* pq = incgam(h,pow(psiCav/d,c));
       double Epsimin = kxylemmax*(-d/c)*tgamma(h)*pq[0];
+      delete[] pq;
       E = E - Epsimin + xylemConductance(psiCav,kxylemmax,c,d)*(-psiCav); //Remove part of the integral corresponding to psimin and add square integral
     }
   }